refactor : 초기화 코드 커스텀 훅으로 분리
